% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_tzar_main.R
\name{local_build_parameters_list}
\alias{local_build_parameters_list}
\title{Dummy function to build parameters list if tzar was not run.}
\usage{
local_build_parameters_list()
}
\value{
list of parameters corresponding to what would have been in tzar's
  project.yaml file
}
\description{
This function is just a place holder that other users could modify
if they were neither running tzar nor running the tzar emulator.
They would just change all of the NA values in here to whatever
they want for the run.
More lines may need to be added to this in the future if more
values from the parameters list are used in this program.
I generated this list by running the following bash command and
editing the resulting lines in the old bdprobdiff project:
     grep 'parameters\$' generateSetCoverProblem.R
}
\details{
Note that the variables listed in this routine are very likely to be
out of date due to the constant evolution of the rest of the code
for this project.  If you're going to use this local_build_parameters_list()
function, you should rebuild it to match the variables currently used in the
rest of the code.  The code shown here for this function is only here
to provide an example of what it might look like based on the parameters
that were in use when it was first generated.
}
\examples{
\dontrun{
local_build_parameters_list ()
}
}
