% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_tzar_main.R
\name{run_without_tzar_or_tzar_emulation}
\alias{run_without_tzar_or_tzar_emulation}
\title{Example function for running the main code without using tzar or tzar
emulation}
\usage{
run_without_tzar_or_tzar_emulation()
}
\description{
All experiments for the paper described in this package are done under tzar
or tzar emulation.  However, run_without_tzar_or_tzar_emulation() is provided
to document how you would call the code for the paper without using tzar.
}
\details{
In that case, you would have to create the required parameters list yourself.
An example function for doing that is provided in get_parameters().

Neither non_tzar_main() nor get_parameters() are used in creating the paper
described in this package, but both functions are provided here as examples.
}
\examples{
\dontrun{
run_without_tzar_or_tzar_emulation (parameters)
}
}
