% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzar_main.R
\name{runt}
\alias{runt}
\title{Convenience function to run tzar with all appropriate variables set}
\usage{
runt()
}
\value{
results of running tzar
}
\description{
Convenience function to run tzar with all appropriate variables set
}
\examples{
\dontrun{
runt ()
}
}

