% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzar_main.R
\name{runtop}
\alias{runtop}
\title{Convenience function to run tzar outside pkg with all appropriate variables set}
\usage{
runtop()
}
\value{
results of running tzar
}
\description{
The name stands for "run Tzar Outside Package".  The function is intended to
be quick and easy to call over and over from the command line, so its
name is short and the call has no arguments.
You have to modify the arguments
set inside the function to correspond to your situation.
}
\examples{
\dontrun{
runtop ()
}
}
