% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSetCoverProblem.R
\name{xu_paper_main}
\alias{xu_paper_main}
\title{Mainline for experiments and data generation for Xu bdpg paper}
\usage{
xu_paper_main(parameters)
}
\arguments{
\item{parameters}{List of parameters controlling the current run (usually
decoded from project.yaml by tzar)}
}
\value{
RETURN_DESCRIPTION
}
\description{
This is the top level function for running experiments and generating data for
the problem difficulty paper that explains the Xu biodiversity problem
generator.
}
\examples{
\dontrun{
xu_paper_main (parameters)
}
}

