% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzar_main.R
\name{runt}
\alias{runt}
\title{Convenience function to call run_tzar without having to enter arguments}
\usage{
runt()
}
\arguments{
\item{main_function}{name of function to call to run under tzar or tzar emulation}

\item{parameters_yaml_file_path}{Full path with file name for the
project.yaml file}

\item{tzar_emulation_yaml_file_path}{Full path with file name for the
yaml file containing control values for tzar emulation}
}
\value{
Returns nothing
}
\description{
Convenience function to call run_tzar without having to enter arguments
}
\examples{
\dontrun{
runt ()
}
}
