% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzar_main.R
\name{runt}
\alias{runt}
\title{Convenience function to call run_tzar without having to enter arguments}
\usage{
runt()
}
\value{
Returns nothing
}
\description{
Convenience function to call run_tzar without having to enter arguments
}
\examples{
\dontrun{
runt ()
}
}
