% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzar_main.R
\name{tzar_main}
\alias{tzar_main}
\title{Wrapper function to call your application code from tzar}
\usage{
tzar_main(parameters, emulating_tzar = FALSE)
}
\arguments{
\item{parameters}{List of parameters controlling the current run (usually
decoded from project.yaml by tzar)}
}
\value{
Returns nothing
}
\description{
Wrapper function to call your application code from tzar
}
\examples{
\dontrun{
tzar_main (parameters)
}
}
