% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tzar_main.R
\name{tzar_main}
\alias{tzar_main}
\title{Wrapper function to call your application code from tzar}
\usage{
tzar_main(parameters)
}
\arguments{
\item{parameters}{list of parameters built by tzar from project.yaml}
}
\description{
Wrapper function to call your application code from tzar
}
\examples{
\dontrun{
tzar_main (parameters)
}
}
